Unit tests for decimals added to Wallet module.

+++ tests/Modules/Finance/Wallet/Unit/TestDoubles/Service/MoneyAmountNormalizer.php
@@ -6,7 +6,7 @@ namespace App\Tests\Modules\Finance\Wallet\Unit\TestDoubles\Service;
 
 final class MoneyAmountNormalizer
 {
-    public static function normalize(int $amount): int
+    public static function normalize(int $amount): float
     {
         return $amount / 100;
     }
